+++ packages/evm/contracts/incentives/AERC20VariableCriteriaIncentive.sol
@@ -24,7 +24,7 @@ abstract contract AERC20VariableCriteriaIncentive is ERC20VariableIncentive {
     struct IncentiveCriteria {
         SignatureType criteriaType;
         bytes32 signature;
-        uint8 fieldIndex;
+        uint32 fieldIndex;
         address targetContract;
     }
